+++ packages/api/src/repository.js
@@ -55,7 +55,7 @@ class Repository {
     return data.ticket;
   }
 
-  async allowedToPerformTicket (action, permission) {
+  allowedToPerformTicket (action, permission) {
     if (action === 'delete-repo') {
       return permission === 'admin';
     }
@@ -74,7 +74,7 @@ class Repository {
       if (ticket.action === 'delete-test') {
         deleter.deleteTest(ticket.fullName, ticket.testName, client);
       } else if (ticket.action === 'delete-repo') {
-        deleter.deleteRepo(client, ticket.fullName);
+        deleter.deleteRepo(ticket.fullName, client);
       } else {
         throw Error('invalid action');
       }