+++ packages/frontend/src/app/repository/tests-list/test-details/test-details.component.ts
@@ -14,6 +14,9 @@
 
 import {Component, Input} from '@angular/core';
 import {Test} from 'src/app/services/search/interfaces';
+import {COMService} from '../../../services/com/com.service';
+import {environment} from '../../../../environments/environment';
+import {RouteProvider} from 'src/app/routing/route-provider/RouteProvider';
 
 @Component({
   selector: 'app-test-details',
@@ -21,10 +24,31 @@ import {Test} from 'src/app/services/search/interfaces';
   styleUrls: ['./test-details.component.css'],
 })
 export class TestDetailsComponent {
+  constructor(public comService: COMService) {}
   @Input() test: Test;
+  @Input() repoName: string;
+  @Input() orgName: string;
+  @Input() removalButtonState;
+  windowProvider = window;
 
-  //Converts tests' passing percentage from decimal to percentage
+  // Converts tests' passing percentage from decimal to percentage
   toPercentage(percentpassing: number): string {
     return (percentpassing * 100).toFixed(2);
+  }
+
+  startDeleteTest() {
+    this.removalButtonState.disabled = true;
+    this.comService
+      .fetchDeleteTestUrl(
+        this.orgName,
+        this.repoName,
+        this.test.name,
+        environment.baseUrl +
+          '/' +
+          RouteProvider.routes.repo.link(this.orgName, this.repoName)
+      )
+      .subscribe(res => {
+        this.windowProvider.open(res, '_self');
+      });
   }
 }