+++ packages/frontend/src/app/services/com/api.ts
@@ -17,9 +17,17 @@ import {environment} from '../../../environments/environment';
 interface ApiLinks {
   get: {
     repositories: (orgName: string) => string;
-    builds: (repoName: string, orgName: string) => string;
+    batches: (repoName: string, orgName: string) => string;
+    batch: (repoName: string, orgName: string, timestamp: number) => string;
     tests: (repoName: string, orgName: string) => string;
     repository: (repoName: string, orgName: string) => string;
+    deleteTest: (
+      orgName: string,
+      repoName: string,
+      testName: string,
+      redirect: string
+    ) => string;
+    deleteRepo: (orgName: string, repoName: string, redirect: string) => string;
   };
   post: {
     authLink: string;
@@ -30,12 +38,37 @@ export const apiBaseLink = environment.baseUrl + '/api/';
 export const apiLinks: ApiLinks = {
   get: {
     repositories: (orgName: string) => apiBaseLink + 'org/' + orgName,
-    builds: (repoName: string, orgName: string) =>
-      apiBaseLink + 'repo/' + orgName + '/' + repoName + '/builds',
+    batches: (repoName: string, orgName: string) =>
+      apiBaseLink + 'repo/' + orgName + '/' + repoName + '/batches',
+    batch: (repoName: string, orgName: string, timestamp: number) =>
+      apiBaseLink + 'repo/' + orgName + '/' + repoName + '/batch/' + timestamp,
     tests: (repoName: string, orgName: string) =>
       apiBaseLink + 'repo/' + orgName + '/' + repoName + '/tests',
     repository: (repoName: string, orgName: string) =>
       apiBaseLink + 'repo/' + orgName + '/' + repoName,
+    deleteTest: (
+      orgName: string,
+      repoName: string,
+      testName: string,
+      redirect: string
+    ) =>
+      apiBaseLink +
+      'repo/' +
+      encodeURIComponent(orgName) +
+      '/' +
+      encodeURIComponent(repoName) +
+      '/test/deleteurl?testname=' +
+      encodeURIComponent(testName) +
+      '&redirect=' +
+      encodeURIComponent(redirect),
+    deleteRepo: (orgName: string, repoName: string, redirect: string) =>
+      apiBaseLink +
+      'repo/' +
+      encodeURIComponent(orgName) +
+      '/' +
+      encodeURIComponent(repoName) +
+      '/deleteurl?redirect=' +
+      encodeURIComponent(redirect),
   },
   post: {
     authLink: apiBaseLink + 'auth',