+++ packages/frontend/src/app/services/com/com.service.ts
@@ -16,22 +16,18 @@ import {Injectable} from '@angular/core';
 import {HttpClient, HttpParams, HttpErrorResponse} from '@angular/common/http';
 import {apiLinks} from './api';
 import {Observable, empty, throwError} from 'rxjs';
-import {
-  Repository,
-  ApiRepository,
-  Filter,
-  Tests,
-  ApiRepositories,
-} from '../search/interfaces';
+import {Repository, Filter, Tests, ApiRepositories} from '../search/interfaces';
 import {catchError} from 'rxjs/operators';
 import {NotFoundError} from './Errors/NotFoundError';
 import {SnackBarService} from '../snackbar/snack-bar.service';
+import * as moment from 'moment';
+import {BuildBatch, Build} from 'src/app/heat-map/services/interfaces';
 
 @Injectable({
   providedIn: 'root',
 })
 export class COMService {
-  constructor(private http: HttpClient, private snackBar: SnackBarService) {}
+  constructor(private http: HttpClient, public snackBar: SnackBarService) {}
 
   public fetchRepositories(
     repoName: string,
@@ -47,14 +43,28 @@ export class COMService {
       .pipe(catchError(err => this.handleError(err)));
   }
 
-  public fetchBuilds(
+  public fetchBatches(
     repoName: string,
     orgName: string,
     filters: Filter[]
-  ): Observable<ApiRepository> {
-    filters.push({name: 'limit', value: '999999'});
+  ): Observable<BuildBatch[]> {
+    filters.push({name: 'utcOffset', value: moment().utcOffset().toString()});
     return this.http
-      .get<ApiRepository>(apiLinks.get.builds(repoName, orgName), {
+      .get<BuildBatch[]>(apiLinks.get.batches(repoName, orgName), {
+        params: this.getParams(filters),
+      })
+      .pipe(catchError(err => this.handleError(err)));
+  }
+
+  public fetchBatch(
+    repoName: string,
+    orgName: string,
+    timestamp: number,
+    filters: Filter[]
+  ): Observable<Build[]> {
+    filters.push({name: 'utcOffset', value: moment().utcOffset().toString()});
+    return this.http
+      .get<Build[]>(apiLinks.get.batch(repoName, orgName, timestamp), {
         params: this.getParams(filters),
       })
       .pipe(catchError(err => this.handleError(err)));
@@ -80,6 +90,31 @@ export class COMService {
     return this.http
       .get<Repository>(apiLinks.get.repository(repoName, orgName), {
         params: this.getParams(filters),
+      })
+      .pipe(catchError(err => this.handleError(err)));
+  }
+
+  public fetchDeleteTestUrl(
+    orgName: string,
+    repoName: string,
+    testName: string,
+    redirect: string
+  ): Observable<string> {
+    return this.http
+      .get(apiLinks.get.deleteTest(orgName, repoName, testName, redirect), {
+        responseType: 'text',
+      })
+      .pipe(catchError(err => this.handleError(err)));
+  }
+
+  public fetchDeleteRepoUrl(
+    orgName: string,
+    repoName: string,
+    redirect: string
+  ): Observable<string> {
+    return this.http
+      .get(apiLinks.get.deleteRepo(orgName, repoName, redirect), {
+        responseType: 'text',
       })
       .pipe(catchError(err => this.handleError(err)));
   }