+++ packages/action/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+### [1.0.3](https://www.github.com/GoogleCloudPlatform/flaky-service/compare/v1.0.2...v1.0.3) (2020-08-17)
+
+
+### Bug Fixes
+
+* Github Action handles endpoints ending in / ([#249](https://www.github.com/GoogleCloudPlatform/flaky-service/issues/249)) ([78b0d7a](https://www.github.com/GoogleCloudPlatform/flaky-service/commit/78b0d7af16e002c9f14e2f0a8a84d00f1b7a696c))
+
+### [1.0.2](https://www.github.com/GoogleCloudPlatform/flaky-service/compare/v1.0.1...v1.0.2) (2020-08-14)
+
+
+### Bug Fixes
+
+* default endpoint for action is production ([#234](https://www.github.com/GoogleCloudPlatform/flaky-service/issues/234)) ([043c602](https://www.github.com/GoogleCloudPlatform/flaky-service/commit/043c60247fe53c995dcbbc7bddbd746d089d8b0f))
+
 ### [1.0.1](https://www.github.com/GoogleCloudPlatform/flaky-service/compare/v1.0.0...v1.0.1) (2020-08-07)